create router files and a basic error page
