+++ routes/quiz.js
@@ -1,0 +1,22 @@
+/*
+ * All routes for /quiz/... are defined here
+ * See README for a list of routes.
+ */
+
+const quizRouter = (db) => {
+  router.get("/", (req, res) => {
+    db.query(`SELECT * FROM users;`)
+      .then(data => {
+        const users = data.rows;
+        res.json({ users });
+      })
+      .catch(err => {
+        res
+          .status(500)
+          .json({ error: err.message });
+      });
+  });
+  return router;
+};
+
+module.exports = quizRouter;
