+++ views/error.ejs
@@ -1,0 +1,6 @@
+<!DOCTYPE html>
+<html lang="en">
+  <body>
+    ERROR! 404
+  </body>
+</html>
